core/txdb: delete older snapshots
After saving a new snapshot, delete older snapshots.

We could change the snapshots table to be a singleton table, but the
/get-snapshot RPC uses content-addressable snapshots. There would be
a race condition between when a client looks up the height of the
most recent snapshot and when they retrieve that snapshot.

Also, I'm not sure if there are use cases for keeping more snapshots
than just the most recent snapshot.