core/account: add expiry to control programs
Add a required expiration to all control programs. For control programs
automatically created via a build action, the expiration is the built
transaction's max time. For explicitly created control programs, the
user must specify its expiration. Otherwise, it defaults to 30 days.

The expiration is only considered to be elapsed once the Core ingests a
block with a timestamp greater than the control program's expiration.
This allows us to use a transaction's max time as the control program's
expiry without worrying about clock skew. Once the expiration has
elapsed, the control programs are deleted and forgotten.

If any assets are sent to an expired control program, the Core will not
index them (so they will not be spendable) or annotate them as belonging
to any account.